refactor(billets-web): :art: modified ui
