+++ apps/billets-web/app/(ui)/index.ts
@@ -1,5 +1,7 @@
 export * from './app-footer'
 export * from './app-header'
 export * from './app-layout'
+export * from './network-error'
 export * from './page-layout'
 export * from './page-top'
+export * from './unknown-error'
