+++ apps/billets-web/app/(ui)/network-error/index.ts
