+++ apps/billets-web/app/(ui)/network-error/network-error.styled.ts
