--- apps/billets-web/app/ui/index.ts
@@ -1,2 +1,0 @@
-export * from './network-error'
-export * from './unknown-error'
