+++ apps/billets-web/libs/registries/api-error-boundary-registry/api-error-boundary-registry.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { NetworkError } from 'app/ui'
+import { NetworkError } from 'app/(ui)'
 import { PropsWithChildren } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
