+++ apps/billets-web/libs/registries/global-error-boundary-registry/global-error-boundary-registry.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { UnknownError } from 'app/ui'
+import { UnknownError } from 'app/(ui)'
 import { PropsWithChildren } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
